pyproject.toml
[project]
name = "beget-mcp"
version = "0.1.0"
description = "Beget hosting MCP server — manage sites, domains, MySQL, FTP, DNS, cron, backups via AI agents"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "requests",
    "python-dotenv",
]

[project.scripts]
beget-mcp = "mcp_beget.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_beget"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["mcp_beget"]
